# backend/openapi-spec-v2.yaml
openapi: 3.1.0
info:
  title: Govpulse Backend Service
  version: 2.0.0
  description: |
    This is the backend service for the Govpulse application, providing endpoints for user management, issue tracking, appointment scheduling, and more.
    With v2 it supports Clerk for user authentication and authorization.
...
paths:
...
  /api/v2/upload-image:
    post:
      summary: Upload multiple images
      operationId: uploadMultipleImages
      description: Uploads one or more images to the server and stores them on Cloudinary.
      tags:
        - images-related
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  description: List of image files to upload
                  items:
                    type: string
                    format: binary
              required:
                - images
      responses:
        "200":
          description: Successfully uploaded images
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successfully uploaded 2 images"
                  images:
                    type: array
                    items:
                      type: object
                      properties:
                        success:
                          type: boolean
                          example: true
                        url:
                          type: string
                          format: uri
                          example: "https://res.cloudinary.com/gov-p/image/upload/v1755059486/govpulse/uploads/image_1755059475754_5879461.png"
                        public_id:
                          type: string
                          example: "govpulse/uploads/image_1755059475754_5879461"
                        width:
                          type: integer
                          example: 10197
                        height:
                          type: integer
                          example: 2622
                        format:
                          type: string
                          example: "png"
                        bytes:
                          type: integer
                          example: 2089919
                        created_at:
                          type: string
                          format: date-time
                          example: "2025-08-13T04:31:26Z"
        "400":
          description: Bad Request (e.g., no images provided or some uploads failed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "No images provided"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Internal server error during upload"
...
  /api/v2/time-slots/view-free-time-slots:
    get:
      summary: View free time slots
      operationId: viewFreeTimeSlots
      description: Allows government officials to view available free time slots of their employed Authority. Note that the authority is identified using the bearer token.
      tags:
        - time-slot-related
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.freeTimeSlotData"
        "401":
          description: Unauthorized - Invalid token
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized - Invalid token
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security:
        - BearerAuth: []
...
  /api/v2/time-slots/remove-free-time-slot:
    post:
      summary: Remove free time slot
      operationId: removeFreeTimeSlot
      description: Allows government officials to remove a free time slot for appointments related to their employed Authority. Note that the authority is identified using the bearer token.
      tags:
        - time-slot-related
      requestBody:
        description: Contains the details of the free time slot to be removed
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date-time
                  example: "2025-08-15T00:00:00.000Z"
                start_time:
                  type: string
                  example: "10:00"
                end_time:
                  type: string
                  example: "11:00"
              required:
                - date
                - start_time
                - end_time
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Free time slot removed successfully
                  free_time_slot_data:
                    $ref: "#/components/schemas/models.freeTimeSlotData"
        "400":
          description: Bad Request - Invalid input data
          content:
            text/plain:
              schema:
                type: string
                example: Bad Request - Invalid input data
        "401":
          description: Unauthorized - Invalid token
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized - Invalid token              
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security:
        - BearerAuth: []
  /api/v2/time-slots/get-free-time-slots-of-authority/{authority_id}:
    get:
      summary: Get free time slots of an authority
      operationId: getFreeTimeSlotsOfAnAuthority
      description: Retrieves all free time slots of a specific authority, ment to be used by users.
      tags:
        - time-slot-related
      parameters:
        - name: authority_id
          in: path
          required: true
          description: The ID of the authority whose free time slots are to be retrieved
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.freeTimeSlotData"
        "404":
          description: Not Found - Authority not found or no free time slots available
          content:
            text/plain:
              schema:
                type: string
                example: Not Found - Authority not found or no free time slots available
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    models.freeTimeSlotData:
      type: object
      properties:
        authority_id:
          type: integer
          format: int64
          example: 1
        date:
          type: string
          format: date-time
          example: "2025-08-15T00:00:00.000Z"
        time_slots:
          type: array
          items:
            type: string
            example: "10:00 - 11:00"
          example:
            - "10:00 - 11:00"
            - "11:30 - 12:30"
            - "14:00 - 15:00"
      required:
        - authority_id
        - date
        - time_slots
...